Plots result of signifPlot.

Plots result of signifPlot.

Usage

signifPlot(
  results,
  fdr = FALSE,
  type = "bubble",
  breaks = NULL,
  comparisonGroup = NULL,
  colours = c("#4575B4", "white", "#D73027"),
  marksToPlot = NULL,
  cutoff = 0.05
)

Arguments

results

Data frame obtained from spicy.

fdr

TRUE if FDR correction is used.

type

Where to make a bubble plot or heatmap.

breaks

Vector of 3 numbers giving breaks used in pheatmap. The first number is the minimum, the second is the maximum, the third is the number of breaks.

comparisonGroup

A string specifying the name of the outcome group to compare with the base group.

colours

Vector of colours to use in pheatmap.

marksToPlot

Vector of marks to include in pheatmap.

cutoff

significance threshold for circles in bubble plot

Value

a pheatmap object

Examples

data(spicyTest)

p <- signifPlot(spicyTest, breaks = c(-3, 3, 0.5))
# plot includes unicode characters, do not use default pdf device
ggplot2::ggsave(p, filename = tempfile(), device = cairo_pdf)
#> Saving 6.67 x 6.67 in image