Accessors for SegmentedCells

Methods to access various components of the `SegmentedCells` object.

Usage

cellSummary(x, imageID = NULL, bind = TRUE)

cellSummary(x, imageID = NULL) <- value

cellMarks(x, imageID = NULL, bind = TRUE)

cellMarks(x, imageID = NULL) <- value

cellMorph(x, imageID = NULL, bind = TRUE)

cellMorph(x, imageID = NULL) <- value

imagePheno(x, imageID = NULL, bind = TRUE, expand = FALSE)

imagePheno(x, imageID = NULL) <- value

imageID(x, imageID = NULL)

cellID(x, imageID = NULL)

cellID(x) <- value

imageCellID(x, imageID = NULL)

imageCellID(x) <- value

cellType(x, imageID = NULL)

cellType(x, imageID = NULL) <- value

filterCells(x, select)

cellAnnotation(x, variable, imageID = NULL)

cellAnnotation(x, variable, imageID = NULL) <- value

Arguments

x

A `SegmentedCells` object.

imageID

A vector of imageIDs to specifically extract.

bind

When false outputs a list of DataFrames split by imageID

expand

Used to expand the phenotype information from per image to per cell.

value

The relevant information used to replace.

select

A logical vector of the cells to be kept.

variable

A variable to add or retrieve from cellSummary.

Value

DataFrame or a list of DataFrames

Descriptions

`cellSummary`:

Retrieves the DataFrame containing `x` and `y` coordinates of each cell as well as `cellID`, `imageID` and `cellType`. imageID can be used to select specific images and bind=FALSE outputs the information as a list split by imageID.

`cellMorph`:

Retrieves the DataFrame containing morphology information.

`cellMarks`:

Retrieves the DataFrame containing intensity of gene or protein markers.

`imagePheno`:

Retrieves the DataFrame containing the phenotype information for each image. Using expand = TRUE will produce a DataFrame with the number of rows equal to the number of cells.

Examples

### Something that resembles cellProfiler data

set.seed(51773)

n = 10

cells <- data.frame(row.names = seq_len(n))
cells$ObjectNumber <- seq_len(n)
cells$ImageNumber <- rep(1:2,c(n/2,n/2))
cells$AreaShape_Center_X <- runif(n)
cells$AreaShape_Center_Y <- runif(n)
cells$AreaShape_round <- rexp(n)
cells$AreaShape_diameter <- rexp(n, 2)
cells$Intensity_Mean_CD8 <- rexp(n, 10)
cells$Intensity_Mean_CD4 <- rexp(n, 10)

cellExp <- SegmentedCells(cells, cellProfiler = TRUE)
#> There is no cellType column, setting to NA
#> Warning: The dim() method for DataFrameList objects is deprecated. Please use
#>   dims() on these objects instead.
#> Warning: The nrow() method for DataFrameList objects is deprecated. Please use
#>   nrows() on these objects instead.
#> Warning: The ncol() method for CompressedSplitDataFrameList objects is
#>   deprecated. Please use ncols() on these objects instead.
#> Warning: The dim() method for DataFrameList objects is deprecated. Please use
#>   dims() on these objects instead.
#> Warning: The nrow() method for DataFrameList objects is deprecated. Please use
#>   nrows() on these objects instead.
#> Warning: The ncol() method for CompressedSplitDataFrameList objects is
#>   deprecated. Please use ncols() on these objects instead.
#> Warning: The dim() method for DataFrameList objects is deprecated. Please use
#>   dims() on these objects instead.
#> Warning: The nrow() method for DataFrameList objects is deprecated. Please use
#>   nrows() on these objects instead.
#> Warning: The ncol() method for CompressedSplitDataFrameList objects is
#>   deprecated. Please use ncols() on these objects instead.
#> Warning: The dim() method for DataFrameList objects is deprecated. Please use
#>   dims() on these objects instead.
#> Warning: The nrow() method for DataFrameList objects is deprecated. Please use
#>   nrows() on these objects instead.
#> Warning: The ncol() method for CompressedSplitDataFrameList objects is
#>   deprecated. Please use ncols() on these objects instead.

### Cluster cell types
intensities <- cellMarks(cellExp)
kM <- kmeans(intensities,2)
cellType(cellExp) <- paste('cluster',kM$cluster, sep = '')

cellSummary(cellExp, imageID = 1)
#> DataFrame with 5 rows and 6 columns
#>    imageID      cellID imageCellID         x         y cellType
#>   <factor> <character> <character> <numeric> <numeric> <factor>
#> 1        1      cell_1      cell_1  0.367243  0.820150 cluster1
#> 2        1      cell_2      cell_2  0.613818  0.846200 cluster1
#> 3        1      cell_3      cell_3  0.335930  0.318286 cluster2
#> 4        1      cell_4      cell_4  0.501737  0.373931 cluster2
#> 5        1      cell_5      cell_5  0.829350  0.140415 cluster1